+++ src/main/java/net/jewelry/village/JewelryVillagers.java
@@ -15,9 +15,11 @@ import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.util.Identifier;
 import net.minecraft.village.TradeOffer;
+import net.minecraft.village.TradeOffers;
 import net.minecraft.village.VillagerProfession;
 import net.minecraft.world.poi.PointOfInterestType;
 
+import java.util.LinkedHashMap;
 import java.util.List;
 
 public class JewelryVillagers {
@@ -44,35 +46,35 @@ public class JewelryVillagers {
         );
     }
 
-    private static class Offer {
-        int level;
-        ItemStack input;
-        ItemStack output;
-        int maxUses;
-        int experience;
-        float priceMultiplier;
-
-        public Offer(int level, ItemStack input, ItemStack output, int maxUses, int experience, float priceMultiplier) {
-            this.level = level;
-            this.input = input;
-            this.output = output;
-            this.maxUses = maxUses;
-            this.experience = experience;
-            this.priceMultiplier = priceMultiplier;
-        }
-
-        public static Offer buy(int level, ItemStack item, int price, int maxUses, int experience, float priceMultiplier) {
-            return new Offer(level, item, new ItemStack(Items.EMERALD, price), maxUses, experience, priceMultiplier);
-        }
-
-        public static Offer sell(int level, ItemStack item, int price, int maxUses, int experience, float priceMultiplier) {
-            return new Offer(level, new ItemStack(Items.EMERALD, price), item, maxUses, experience, priceMultiplier);
-        }
-    }
+//    private static class Offer {
+//        int level;
+//        ItemStack input;
+//        ItemStack output;
+//        int maxUses;
+//        int experience;
+//        float priceMultiplier;
+//
+//        public Offer(int level, ItemStack input, ItemStack output, int maxUses, int experience, float priceMultiplier) {
+//            this.level = level;
+//            this.input = input;
+//            this.output = output;
+//            this.maxUses = maxUses;
+//            this.experience = experience;
+//            this.priceMultiplier = priceMultiplier;
+//        }
+//
+//        public static Offer buy(int level, ItemStack item, int price, int maxUses, int experience, float priceMultiplier) {
+//            return new Offer(level, item, new ItemStack(Items.EMERALD, price), maxUses, experience, priceMultiplier);
+//        }
+//
+//        public static Offer sell(int level, ItemStack item, int price, int maxUses, int experience, float priceMultiplier) {
+//            return new Offer(level, new ItemStack(Items.EMERALD, price), item, maxUses, experience, priceMultiplier);
+//        }
+//    }
 
     public static void register() {
-        var wizardPOI = registerPOI(JEWELER, JewelryBlocks.JEWELERS_KIT.block());
-        var wizardMerchantProfession = registerProfession(
+        var poi = registerPOI(JEWELER, JewelryBlocks.JEWELERS_KIT.block());
+        var profession = registerProfession(
                 JEWELER,
                 RegistryKey.of(Registries.POINT_OF_INTEREST_TYPE.getKey(), Identifier.of(JewelryMod.ID, JEWELER)));
 
@@ -111,14 +113,44 @@ public class JewelryVillagers {
 //                Offer.sell(5, JewelryItems.tanzanite_necklace.item().getDefaultStack(), 45, 3, 13, 0.1f)
 //        );
 
-//        for(var offer: wizardMerchantOffers) {
-//            TradeOfferHelper.registerVillagerOffers(wizardMerchantProfession, offer.level, factories -> {
-//                factories.add(((entity, random) -> new TradeOffer(
-//                        offer.input,
-//                        offer.output,
-//                        offer.maxUses, offer.experience, offer.priceMultiplier)
-//                ));
-//            });
-//        }
+        LinkedHashMap<Integer, List<TradeOffers.Factory>> trades = new LinkedHashMap<>();
+
+        trades.put(1, List.of(
+                new TradeOffers.BuyItemFactory(Items.COPPER_INGOT, 8, 8, 3, 2),
+                new TradeOffers.BuyItemFactory(Items.STRING, 7, 6, 3, 2),
+                new TradeOffers.SellItemFactory(JewelryItems.copper_ring.item(), 4, 1, 12, 4)
+        ));
+        trades.put(2, List.of(
+                new TradeOffers.BuyItemFactory(Items.GOLD_INGOT, 7, 8, 2, 8),
+                new TradeOffers.SellItemFactory(JewelryItems.iron_ring.item(), 4, 1, 6, 5),
+                new TradeOffers.SellItemFactory(JewelryItems.gold_ring.item(), 18, 1, 6, 5)
+        ));
+        trades.put(3, List.of(
+                new TradeOffers.BuyItemFactory(Items.DIAMOND, 1, 12, 10, 10),
+                new TradeOffers.SellItemFactory(JewelryItems.emerald_necklace.item(), 20, 1, 12, 10),
+                new TradeOffers.SellItemFactory(JewelryItems.diamond_necklace.item(), 25, 1, 12, 10)
+        ));
+        trades.put(4, List.of(
+                new TradeOffers.SellItemFactory(JewelryItems.ruby_ring.item(), 35, 1, 5, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.topaz_ring.item(), 35, 1, 5, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.citrine_ring.item(), 35, 1, 5, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.jade_ring.item(), 35, 1, 5, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.sapphire_ring.item(), 35, 1, 5, 13),
+                new TradeOffers.SellItemFactory(JewelryItems.tanzanite_ring.item(), 35, 1, 5, 13)
+        ));
+        trades.put(5, List.of(
+                new TradeOffers.SellItemFactory(JewelryItems.ruby_necklace.item(), 45, 1, 3, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.topaz_necklace.item(), 45, 1, 3, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.citrine_necklace.item(), 45, 1, 3, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.jade_necklace.item(), 45, 1, 3, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.sapphire_necklace.item(), 45, 1, 3, 15),
+                new TradeOffers.SellItemFactory(JewelryItems.tanzanite_necklace.item(), 45, 1, 3, 15)
+        ));
+
+        for (var entry: trades.entrySet()) {
+            TradeOfferHelper.registerVillagerOffers(profession, entry.getKey(), factories -> {
+                factories.addAll(entry.getValue());
+            });
+        }
     }
 }